+++ mail/mail-common/tests/draft_undo_send.rs
@@ -1,0 +1,146 @@
+use proton_action_queue::queue::ActionError;
+use proton_api_core::consts::Mail;
+use proton_api_core::services::proton::common::{LabelId, UserId};
+use proton_api_core::services::proton::prelude::ApiErrorInfo;
+use proton_api_mail::services::proton::prelude::{MessageFlags, PostCancelSendResponse};
+use proton_core_common::models::ModelExtension;
+use proton_mail_common::actions::draft::UndoSend;
+use proton_mail_common::datatypes::SystemLabelId;
+use proton_mail_common::draft::{Draft, Error};
+use proton_mail_common::models::Message;
+use proton_mail_common::MailContextError;
+use proton_mail_test_utils::message_body::{
+    message_body_test_message_simple, message_body_test_params, message_body_test_user_secret,
+    TEST_USER_ID,
+};
+use proton_mail_test_utils::test_context::MailTestContext;
+
+#[tokio::test]
+async fn draft_undo_send() {
+    let ctx = MailTestContext::with_user_secret_and_user_id(
+        message_body_test_user_secret(),
+        UserId::from(TEST_USER_ID),
+    )
+    .await;
+
+    let params = message_body_test_params();
+
+    let user_ctx = ctx.mail_user_context().await;
+
+    let mut tether = user_ctx.user_stash().connection();
+
+    let mut sent_message = message_body_test_message_simple();
+    sent_message.metadata.label_ids.clear();
+    sent_message.metadata.label_ids.push(LabelId::sent());
+    sent_message.metadata.flags.set(MessageFlags::SENT, true);
+
+    let mut undo_message = sent_message.metadata.clone();
+    undo_message.label_ids.clear();
+    undo_message.label_ids.push(LabelId::drafts());
+    undo_message.flags.set(MessageFlags::SENT, false);
+
+    ctx.setup_user(params.clone()).await;
+    ctx.mock_undo_send(
+        sent_message.metadata.id.clone(),
+        Ok(PostCancelSendResponse {
+            message: undo_message,
+        }),
+    )
+    .await;
+    ctx.catch_all().await;
+    ctx.init_user(user_ctx.clone()).await;
+
+    let mut local_sent_message = Message::from_api_metadata(sent_message.metadata.clone(), &tether)
+        .await
+        .unwrap();
+
+    let tx = tether.transaction().await.unwrap();
+    local_sent_message.save(&tx).await.unwrap();
+    tx.commit().await.unwrap();
+
+    // Queue undo send action - The exposed API uses apply for faster reaction, but we
+    // want to check intermediate state here.
+    user_ctx
+        .with_queue(|queue| queue.queue_action(UndoSend::new(local_sent_message.local_id.unwrap())))
+        .await
+        .unwrap();
+
+    let updated_local_message = Message::find_by_id(local_sent_message.local_id.unwrap(), &tether)
+        .await
+        .unwrap()
+        .unwrap();
+
+    assert!(updated_local_message.label_ids.contains(&LabelId::drafts()));
+    assert!(!updated_local_message.label_ids.contains(&LabelId::sent()));
+    assert!(!updated_local_message
+        .flags
+        .contains(MessageFlags::SENT.into()));
+
+    // flush queue.
+    user_ctx.execute_pending_actions().await.unwrap();
+}
+
+#[tokio::test]
+async fn draft_undo_send_failure() {
+    // Check that the message is put back into the sent folder
+    let ctx = MailTestContext::with_user_secret_and_user_id(
+        message_body_test_user_secret(),
+        UserId::from(TEST_USER_ID),
+    )
+    .await;
+
+    let params = message_body_test_params();
+
+    let user_ctx = ctx.mail_user_context().await;
+
+    let mut tether = user_ctx.user_stash().connection();
+
+    let mut sent_message = message_body_test_message_simple();
+    sent_message.metadata.label_ids.clear();
+    sent_message.metadata.label_ids.push(LabelId::sent());
+    sent_message.metadata.flags.set(MessageFlags::SENT, true);
+
+    ctx.setup_user(params.clone()).await;
+    ctx.mock_undo_send(
+        sent_message.metadata.id.clone(),
+        Err(ApiErrorInfo {
+            code: Mail::MessageSentCanNoLongerBeUndone as u32,
+            error: None,
+            details: None,
+        }),
+    )
+    .await;
+    ctx.catch_all().await;
+    ctx.init_user(user_ctx.clone()).await;
+
+    let mut local_sent_message = Message::from_api_metadata(sent_message.metadata.clone(), &tether)
+        .await
+        .unwrap();
+
+    let tx = tether.transaction().await.unwrap();
+    local_sent_message.save(&tx).await.unwrap();
+    tx.commit().await.unwrap();
+
+    // Que undo send action
+    let Err(err) = user_ctx
+        .with_queue(|queue| Draft::action_undo_send(queue, local_sent_message.local_id.unwrap()))
+        .await
+    else {
+        panic!("Should have failed.")
+    };
+
+    let updated_local_message = Message::find_by_id(local_sent_message.local_id.unwrap(), &tether)
+        .await
+        .unwrap()
+        .unwrap();
+
+    assert!(!updated_local_message.label_ids.contains(&LabelId::drafts()));
+    assert!(updated_local_message.label_ids.contains(&LabelId::sent()));
+    assert!(updated_local_message
+        .flags
+        .contains(MessageFlags::SENT.into()));
+    assert!(matches!(
+        err,
+        ActionError::Action(MailContextError::Draft(Error::SendCanNoLongerBeUndone))
+    ));
+}