+++ mail/mail-uniffi/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   - [ET-1417] `Draft::discard` - Discards a draft from the composer.
   - More logging for the html transformations.
+  - [ET-679] `draft_undo_send` - Cancel sending of a message.
 
 ### Fixed